refactor: components 디렉토리 분류
